requirements.txt
opencv_python==4.7.0.72
pyserial==3.5
